+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@neondatabase/serverless": "^1.0.1",
+    "mybatis-mapper": "^0.8.0",
     "next": "15.4.1",
     "next-auth": "^5.0.0-beta.29",
     "react": "19.1.0",
