+++ src/assets/mapper/test.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="test">
+    <select id="selectTest">
+        SELECT /* selectTest */ TO_CHAR(NOW() AT TIME ZONE #{timezone}, #{template}) AS now
+    </select>
+</mapper>
